#42: Promote archetype concept to storage api 2/2
To prepare for performance optimizations in CompositionPlugin the
concept of archetype and accessing entities and their components through
an archetype needs to be part of the storage api module.

Instead of tracking entities through EntitiesEvent and EntityEvent, the
composition plugin will track archetypes through a new
ArchetypeAddedEvent. Archetypes provide an iterator-like way to iterate
through entities and their components, which will provide better memory
locality for the archetype storage.