Refactor entity relation target deletion handling based on callback
Old implementation had issues when a target was related to the same
entity multiple times, only removing one relation type. Adding an entity
multiple times to the IntBag could have been an option, but the callback
variant is simpler and more performant as there is no need of figuring
out which types should be removed.