#59: Add ArchetypeGraphNode to reduce number of archetypes
With 7fea54f the way archetypes are resolved is with a graph structure.
When resolving an archetype for a set of component types, they are
resolved step-by-step by adding the component types to the current
archetype (starting with the empty archetypes) until all component types
are added.

This might introduce a large number of archetypes that will never
contain entities, but will be iterated upon. To reduce this overhead,
introduce a separate ArchetypeGraphNode that contains a lazy Archetype
field instead.