50: Add non-pooled component set iteration WIP
TODOs:
- refactor to archetypeiterator (composition1, compositiondataimpl)
- remove unused code for component sets and compositions
- non-pooled ArchetypeIterableAccessor means no concurrent iteration
over that very same accessor
  - CompositionData<P> for component sets must not be cached/reused
  - write test that verifies this

The goal is to eliminate the use of pooled MyComponentSetAccessor in
compositions and instead generate a immutable IterableComponentAccessor
that can be fully initialized on composition creation and whenever a new
archetype is offered. The usage of final fields for DataAccessor and
ComponentAccessor hopefully allow for better inlining and performance.

This requires that no ComponentAccessor depends on new instances being
retrieved before usage. Implementations must track all created instances
and update them whenever things change that require such updates, or
that implementations must have access to the required data from a parent
instance, such as with WildcardClassMapper.